+++ src/commonMain/kotlin/dsl/ObjectTypeCreator.kt
@@ -5,7 +5,7 @@ class ObjectTypeCreator(
     private val objectTypeIdCreator: PatternIdCreator,
     private val groupsIdCreator: GroupsIdCreator,
 ) {
-    fun createObjectType(label: String, placeNameCreator: ((placeIndexForType: Int) -> String)?) : ObjectTypeDSL {
+    fun createObjectType(label: String, placeNameCreator: ((placeIndexForType: Long) -> String)?) : ObjectTypeDSL {
         return objectTypesContainer.objectTypes.getOrPut(label) {
             val newId = objectTypeIdCreator.newLabelId()
             val localPlaceNameCreator = placeNameCreator ?: {
